apps/backend/pyproject.toml
[project]
name = "embedded-statistics-backend"
version = "1.0.0"
description = "FastAPI backend for embedded statistics tracking"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "motor>=3.3.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

